services/preprocessing_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
pymupdf==1.22.0
pillow==10.0.0
opencv-python==4.7.0.72
scikit-image==0.21.0
numpy==1.26.0
requests==2.31.0
minio==7.2.7
boto3
torch
transformers